rapikan utils
