+++ utils/utils_test.go
@@ -1,0 +1,24 @@
+package utils
+
+import (
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestMain_InputScan(t *testing.T) {
+	str, err := InputScan()
+	assert.Error(t, err, "Harus Error")
+	assert.EqualValues(t, str, "", "kosong")
+}
+
+func TestMain_TimeDateNow_BedaFormat(t *testing.T) {
+	assert.NotEqual(t, TimeDateNow(), time.Now().String(), "Cek tanggal")
+}
+
+func TestMain_TimeDateNow_Ok(t *testing.T) {
+	date := TimeDateNow()
+	strDate := time.Now().Format("2006-01-02 15:04:05")
+	assert.Equal(t, date, strDate, "Harus sesuai")
+}
